Refactored logic and created a FrameCreator class.
